moved evaluation/visualization to another file
